% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{formatReportAllotment}
\alias{formatReportAllotment}
\title{format df in preparation for dplyr (future)selection}
\usage{
formatReportAllotment(df, AllotmentCat, AllotmentCol, unAllotmentCol)
}
\arguments{
\item{df}{dplyr compatabile input type}

\item{AllotmentCol}{'issue_date" xor 'maturity_date"}

\item{unAllotmentCol}{'maturity_date' xor 'issue_date"}

\item{AllotCat}{alotment category point of view, either issue "ISS" or maturity "MTR"}
}
\description{
format df in preparation for dplyr (future)selection
}
\examples{
\dontrun{
NEED one

}
}
