% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{formatTreasuriesAllotmentDateCols}
\alias{formatTreasuriesAllotmentDateCols}
\title{format Investor Class Auction Allotments date-ish columns}
\usage{
formatTreasuriesAllotmentDateCols(treasuriesAllot = NULL, cols = NULL)
}
\arguments{
\item{treasuriesAllot}{Auction Allotments}

\item{cols}{columns in a character vector, to change to type Date}
}
\description{
change data columns to be of  R class Date
}
\examples{
\dontrun{
formatted <- formatTreasuriesAllotmentDateCols(
                data.frame(alpha=1:2), "alpha"
              )
formatted
}
}
