% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSymbols.treasury}
\alias{getSymbols.treasury}
\title{Aquire Bills and Bonds allotment data from the US treasury}
\usage{
getSymbols.treasury(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{One or more treasury 'issue' or 'maturity' bills/bonds symbols}

\item{env}{possibly returned data is loaded into this environment}

\item{return.class}{call of the object sent back to the user}
}
\description{
Aquire Bills and Bonds allotment data from the US treasury
}
\examples{
\dontrun{
library(quantmod)

## issue dates
treasuryTRISS10YEARNOTE <- getSymbols("TRISS10YEARNOTE", src = "treasury", auto.assign = FALSE)
head(treasuryTRISS10YEARNOTE)

## maturity dates
treasuryTRMTR10YEARNOTE <- getSymbols("TRMTR10YEARNOTE", src = "treasury", auto.assign = FALSE)
head(treasuryTRMTR10YEARNOTE)

}
}
