% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{historicalRawTreasuriesAllotment}
\alias{historicalRawTreasuriesAllotment}
\title{Get one historical Investor Class Auction Allotment}
\usage{
historicalRawTreasuriesAllotment(treasClassUrl = NULL)
}
\arguments{
\item{treasClassUrl}{url of one historical allotment}
}
\description{
Get one historical Investor Class Auction Allotment
}
\examples{
\dontrun{
 historical <- historicalRawTreasuriesAllotment()
 historical
}
}
