% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{historicalRawTreasuriesAllotments}
\alias{historicalRawTreasuriesAllotments}
\title{Get historical Investor Class Auction Allotments}
\usage{
historicalRawTreasuriesAllotments(treasList = NULL,
  reorgRawTreasuriesAllotmentFn = NULL)
}
\arguments{
\item{treasList}{list() containting urls of historical allotments}
}
\description{
Get historical Investor Class Auction Allotments
}
\examples{
\dontrun{
 rawHistoricals <- historicalRawTreasuriesAllotments()
 rawHistoricals
}
}
