% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{quantmod___convert.time.series}
\alias{quantmod___convert.time.series}
\title{Quantmod internal conversion}
\usage{
quantmod___convert.time.series(fr, return.class)
}
\arguments{
\item{fr}{data.frame of information: left most column is the 'time' column}

\item{return.class}{call of the object sent back to the user}
}
\description{
Quantmod internal conversion
}
