% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rawTreasuriesAllotments}
\alias{rawTreasuriesAllotments}
\title{Get Investor Class Auction Allotments}
\usage{
rawTreasuriesAllotments(treasuriesAllotments = NULL)
}
\arguments{
\item{treasuriesAllotments}{character vector of function names}
}
\description{
Get both historical and recent Investor Class Auction Allotments
}
\examples{
\dontrun{
rawTreasuries <- rawTreasuriesAllotments()
rawTreasuries
}
}
