% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recentRawTreasuriesAllotment}
\alias{recentRawTreasuriesAllotment}
\title{Get one recent Investor Class Auction Allotment}
\usage{
recentRawTreasuriesAllotment(treasClassUrl = NULL)
}
\arguments{
\item{treasClassUrl}{url of one single recent allotment}
}
\description{
Get one recent Investor Class Auction Allotment
}
\examples{
\dontrun{
recentRaw <- recentRawTreasuriesAllotment()
recentRaw
}
}
