% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recentRawTreasuriesAllotments}
\alias{recentRawTreasuriesAllotments}
\title{Get many recent Investor Class Auction Allotments}
\usage{
recentRawTreasuriesAllotments(treasList = NULL)
}
\arguments{
\item{treasList}{list() containting urls of recent allotments}
}
\description{
Get many recent Investor Class Auction Allotments
}
\examples{
\dontrun{
recents <- recentRawTreasuriesAllotment()
recents
}
}
