% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{reorgRawTreasuriesAllotment}
\alias{reorgRawTreasuriesAllotment}
\title{reorganize Investor Class Auction Allotments}
\usage{
reorgRawTreasuriesAllotment(treasAllotment = NULL)
}
\arguments{
\item{treasAllotment}{treasury allotmetn data.frame to be reorganized}
}
\description{
for 'old' Class Auction Allotments, change 'security_type' and/or 'security_term'
to be just (condensed to one column) 'security' with proper naming
compatible with the 'new'(er) Class Auction Allotments
}
\examples{
\dontrun{
reorgedTreasury1 <-  reorgRawTreasuriesAllotment(
               data.frame(security_term = c("BOND", "NOTE", "YR")
             , security_type = rep("",3)
             , issue_date = rep(NA,3)
             , stringsAsFactors = FALSE)
           )
reorgedTreasury1

reorgedTreasury2 <-  reorgRawTreasuriesAllotment(
               data.frame(security_term = c("CASH", "WK", "YR")
             , issue_date = rep(NA,3)
             , stringsAsFactors = FALSE)
           )
reorgedTreasury2
}
}
